js rename to mjs
