+++ acorn-loose/src/parseutil.mjs
