+++ acorn-walk/src/index.mjs
