+++ acorn/rollup.config.bin.js
@@ -1,7 +1,7 @@
 import buble from "rollup-plugin-buble"
 
 export default {
-  entry: "acorn/src/bin/acorn.js",
+  entry: "acorn/src/bin/acorn.mjs",
   dest: "acorn/dist/bin.js",
   format: "cjs",
   external: ["fs", "path", "acorn"],
