+++ acorn/src/index.mjs
@@ -14,23 +14,23 @@
 //
 // [walk]: util/walk.js
 
-import {Parser} from "./state"
-import "./parseutil"
-import "./statement"
-import "./lval"
-import "./expression"
-import "./location"
-import "./scope"
+import {Parser} from "./state.mjs"
+import "./parseutil.mjs"
+import "./statement.mjs"
+import "./lval.mjs"
+import "./expression.mjs"
+import "./location.mjs"
+import "./scope.mjs"
 
-export {Parser} from "./state"
-export {defaultOptions} from "./options"
-export {Position, SourceLocation, getLineInfo} from "./locutil"
-export {Node} from "./node"
-export {TokenType, types as tokTypes, keywords as keywordTypes} from "./tokentype"
-export {TokContext, types as tokContexts} from "./tokencontext"
-export {isIdentifierChar, isIdentifierStart} from "./identifier"
-export {Token} from "./tokenize"
-export {isNewLine, lineBreak, lineBreakG, nonASCIIwhitespace} from "./whitespace"
+export {Parser} from "./state.mjs"
+export {defaultOptions} from "./options.mjs"
+export {Position, SourceLocation, getLineInfo} from "./locutil.mjs"
+export {Node} from "./node.mjs"
+export {TokenType, types as tokTypes, keywords as keywordTypes} from "./tokentype.mjs"
+export {TokContext, types as tokContexts} from "./tokencontext.mjs"
+export {isIdentifierChar, isIdentifierStart} from "./identifier.mjs"
+export {Token} from "./tokenize.mjs"
+export {isNewLine, lineBreak, lineBreakG, nonASCIIwhitespace} from "./whitespace.mjs"
 
 export const version = "6.0.1"
