+++ acorn/src/scopeflags.mjs
