+++ acorn/src/state.mjs
@@ -1,8 +1,8 @@
-import {reservedWords, keywords} from "./identifier"
-import {types as tt} from "./tokentype"
-import {lineBreak} from "./whitespace"
-import {getOptions} from "./options"
-import {SCOPE_TOP, SCOPE_FUNCTION, SCOPE_ASYNC, SCOPE_GENERATOR} from "./scopeflags"
+import {reservedWords, keywords} from "./identifier.mjs"
+import {types as tt} from "./tokentype.mjs"
+import {lineBreak} from "./whitespace.mjs"
+import {getOptions} from "./options.mjs"
+import {SCOPE_TOP, SCOPE_FUNCTION, SCOPE_ASYNC, SCOPE_GENERATOR} from "./scopeflags.mjs"
 
 function keywordRegexp(words) {
   return new RegExp("^(?:" + words.replace(/ /g, "|") + ")$")
