+++ acorn/src/whitespace.mjs
