+++ bin/run_test262.js
@@ -5,9 +5,13 @@ const parse = require("../acorn").parse
 
 const unsupportedFeatures = [
   "BigInt",
-  "class-fields",
   "class-fields-private",
   "class-fields-public",
+  "class-methods-private",
+  "class-static-fields-private",
+  "class-static-fields-public",
+  "class-static-methods-private",
+  "export-star-as-namespace-from-module",
   "numeric-separator-literal"
 ];
