+++ package.json
@@ -34,15 +34,15 @@
     "lint": "eslint acorn/src/ acorn-walk/src/ acorn-loose/src/"
   },
   "devDependencies": {
-    "eslint": "^4.10.0",
+    "eslint": "^4.19.1",
     "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^3.0.1",
     "rollup": "^0.45.0",
     "rollup-plugin-buble": "^0.16.0",
-    "test262": "git+https://github.com/tc39/test262.git#3bfad28cc302fd4455badcfcbca7c5bb7ce41a72",
+    "test262": "git+https://github.com/tc39/test262.git#e286bfa00086226f781a3ed4e0a6295634b8ed11",
     "test262-parser-runner": "^0.4.0",
     "unicode-11.0.0": "^0.7.7"
   }
